Refactor AdminDomainStatsReport
